pyproject.toml
[project]
name = "stafford-gpt-service"
version = "0.1.0"
description = "Stafford Global's intelligent student advisor with canonical data + vector evidence architecture"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "langchain>=0.2.14",
    "langchain-openai>=0.1.23",
    "langchain-community>=0.2.14",
    "openai>=1.55.3",
    "playwright>=1.40.0",
    "lxml>=4.9.3",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "backoff>=2.2.1",
    "docx2txt>=0.9",
    "pandas>=2.2.0",
    "tiktoken>=0.7.0",
    "unstructured>=0.12.0",
    "python-multipart>=0.0.6",
    "pypdf2>=3.0.1",
    "pyee>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
